explicitly install flamegraph tools

+++ scripts/log.ps1
@@ -143,9 +143,13 @@ if ($IsLinux) {
     catch {
         Write-Host "Installing perf"
         sudo apt-get install -y linux-tools-$(uname -r)
+    }
+    if (Test-Path -Path /usr/bin/stackcollapse-perf.pl) {
         wget https://raw.githubusercontent.com/brendangregg/FlameGraph/master/stackcollapse-perf.pl -O /usr/bin/stackcollapse-perf.pl
-        wget https://raw.githubusercontent.com/brendangregg/FlameGraph/master/flamegraph.pl -O /usr/bin/flamegraph.pl
         chmod +x /usr/bin/stackcollapse-perf.pl
+    }
+    if (Test-Path -Path /usr/bin/flamegraph.pl) {
+        wget https://raw.githubusercontent.com/brendangregg/FlameGraph/master/flamegraph.pl -O /usr/bin/flamegraph.pl
         chmod +x /usr/bin/flamegraph.pl
     }
     #Write-Host "Checking perf command...... Done"
